requirements.txt
astroid==2.1.0
baostock==0.8.8
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.3.1
mccabe==0.6.1
msgpack==0.6.1
mysql-connector-python==8.0.15
numpy==1.16.1
pandas==0.24.1
protobuf==3.6.1
pylint==2.2.2
python-dateutil==2.8.0
pytz==2018.9
pyzmq==17.1.2
requests==2.21.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
tushare==1.2.21
urllib3==1.24.1
wrapt==1.11.1
